Corrección base de datos tv_shows

+++ tv_shows/models.py
@@ -3,7 +3,7 @@ from django.db import models
 class Tv_shows(models.Model):
     name = models.CharField(max_length=40)
     price = models.FloatField()
-    year = models.FloatField()
+    year = models.IntegerField()
     director = models.CharField(max_length=40)
     actors = models.CharField(max_length=80)
     description = models.CharField(max_length=300)
